requirements/prod.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
coverage==7.6.8
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.6
filelock==3.16.1
flake8==7.1.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
identify==2.6.3
idna==3.10
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.29.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.5.5
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.26.4
packaging==24.2
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.23.0
plotly-express==0.4.1
pluggy==1.5.0
pre-commit==2.21.0
prompt_toolkit==3.0.48
protobuf==5.28.0
psutil==6.1.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
rich==13.8.0
rich-toolkit==0.11.3
rpds-py==0.20.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.2
streamlit==1.40.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
